refactor: SignStatusController 제거 및 SignDataController로 통합 (#38)
- SignStatusController 삭제
  - 학습 상태 업데이트 관련 기존 API 메서드 이동
  - /api/signs 경로 API 제거
- 새로운 컨트롤러(SignDataController) 추가
  - 외부 API 데이터 정제 및 로드 관련 메서드 추가
  - 학습 상태 업데이트 API 통합
  - /api/sign-data 경로 API 제공
- SignApiService 개선
  - 외부 API 데이터를 'sign_api' -> 'sign' 테이블로 이전하며 저장된 데이터 개수 반환하도록 변경
- 새로운 응답 DTO(SignDataLoadResponseDto) 생성
  - 외부 API 데이터 로드 후 로드된 데이터 수와 메시지 응답
